backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[dotenv]
pandas
numpy
scikit-learn
pdfplumber
python-docx
pytesseract
pillow
openpyxl
xlrd
transformers
torch
prophet
aiofiles
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
boto3
requests